almost finishing the project

+++ app/Http/Controllers/uploadController.php
@@ -26,33 +26,77 @@ class uploadController extends Controller
             "convenable" => $request->convenable,
             "moiyen" => $request->moiyen,
         ];
+        
+        Validator::validate(
+            $input,
+            [
+                "file_import" => ["required", "mimes:xlsx,xls"],
+                "moiyen" => ["required",],
+                "convenable" => ["required"],
+            ],
+            [
 
-        Validator::validate($input, [
-            "file_import" => ["required", "mimes:xlsx,xls"],
-            "moiyen" => ["required"],
-            "convenable" => ["required"],
-        ]);
+                "moiyen" => [
+                    "required" => "moyenne est requis"
+                    
+                ],
+                "convenable" => [
+                    "required" => "convenable est requis"
+                ],
+                "file_import" => [
+                    "required" => "assurez-vous d'importer le fichier",
+                    "mimes" => "le fichier doit être de type:xlsx,xls"
+                ]
+
+            ]
+        );
+
+        if ($request->convenable < $request->moiyen) {
+            return back()->with("error", "convenable devrait être plus grand que moyenne");
+        }
 
         if (DataTable::all()->first()) {
 
             DataTable::truncate();
         }
 
-
-        Excel::import(new dataImport, $request->file("file_import")->store("temp"));
-        $operationOBJ = new OperationController();
-
-        $allFilier = Filiere::all();
-        foreach ($allFilier as $filier) {
-
-            $operationOBJ->getNmbreTotalGroup($filier->code_filiere, $filier->annee);
-            $operationOBJ->getNombreTotalGroupesValides($filier->code_filiere, $filier->annee);
-            $operationOBJ->getNombreTotalGroupesTaux($filier->code_filiere, $filier->annee, $request->convenable, $request->moiyen);
-            $operationOBJ->getTotalModule($filier->code_filiere, $filier->annee);
-            $operationOBJ->getTotalModuleAchever($filier->code_filiere, $filier->annee);
-            $operationOBJ->getTotalEFM_local_regional($filier->code_filiere, $filier->annee);
+        try {
+            //code...
+            Excel::import(new dataImport, $request->file("file_import")->store("temp"));
+        } catch (\Throwable $th) {
+            return back()->with("error", "la forme du fichier n'est pas correcte,le fichier doit contenir ces colonnes (
+                Code Filière,
+                Année de formation,
+                Module,
+                Taux Réalisation (P & SYN ),
+                MH Réalisée Globale,
+                Groupe,
+                Régional,
+                Séance EFM,
+                MH Affectée Globale (P & SYN)
+            )");
         }
 
-       return redirect("/");
+        try {
+            //code...
+            $operationOBJ = new OperationController();
+            $allFilier = Filiere::all();
+            foreach ($allFilier as $filier) {
+
+                $operationOBJ->setDefaultValuesForOtherElements($filier->code_filiere, $filier->annee);
+                $operationOBJ->setDefaultCommentForOtherElements($filier->code_filiere, $filier->annee);
+                $operationOBJ->getNmbreTotalGroup($filier->code_filiere, $filier->annee);
+                $operationOBJ->getNombreTotalGroupesValides($filier->code_filiere, $filier->annee);
+                $operationOBJ->getNombreTotalGroupesTaux($filier->code_filiere, $filier->annee, $request->convenable, $request->moiyen);
+                $operationOBJ->getTotalModule($filier->code_filiere, $filier->annee);
+                $operationOBJ->getTotalModuleAchever($filier->code_filiere, $filier->annee);
+                $operationOBJ->getTotalEFM_local_regional($filier->code_filiere, $filier->annee);
+            }
+            return redirect("/");
+        } catch (\Throwable $th) {
+            //throw $th;
+
+            return back()->with("error", "quelque chose s'est mal passé");
+        }
     }
 }
